Move Work cmdlets to Team
